Remove Protobuf configuration templates
