+++ internal/configuration/schema/configuration.go
@@ -1,0 +1,5 @@
+package schema
+
+type Configuration struct {
+	Hosts []Host `yaml:"hosts"`
+}
