+++ internal/configuration/schema/host.go
