--- internal/configuration/templates/auth.go
@@ -1,3 +1,0 @@
-package templates
-
-const AuthConfigTemplate = ``
