--- internal/configuration/templates/channels.go
@@ -1,88 +1,0 @@
-package templates
-
-const ChannelProfileConfigTemplate = `
-Profile {
-  ProfileId: 0
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-}
-Profile {
-  ProfileId: 1
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-}
-Profile {
-  ProfileId: 2
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-}
-Profile {
-  ProfileId: 3
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-  Channel {
-    ErasureSpecies: "block-4-2"
-    PDiskCategory: 0
-    VDiskCategory: Default
-    StoragePoolKind: "hdd"
-  }
-}
-`
