--- internal/configuration/templates/feature_flag.go
@@ -1,19 +1,0 @@
-package templates
-
-const FeatureFlagsTemplate = `
-EnableSeparateSolomonShardForPDisk: true
-AllowConsistentOperationsForSchemeShard: true
-EnableExternalSubdomains: true
-SendSchemaVersionToDatashard: true
-EnableSchemeBoardCache: true
-EnableSystemViews: true
-EnableExternalHive: true
-UseSchemeBoardCacheForSchemeRequests: true
-CompileMinikqlWithVersion: true
-ReadTableWithSnapshot: true
-EnableOfflineSlaves: true
-AllowOnlineIndexBuild: true
-EnablePersistentQueryStats: true
-DisableDataShardBarrier: true
-EnableTtlOnIndexedTables: true
-`
