--- internal/configuration/templates/init_config_define_storage_pools.go
@@ -1,21 +1,0 @@
-package templates
-
-const DefineStoragePoolsInitConfigTemplate = `
-Command {
-  DefineStoragePool {
-    BoxId: 1
-    StoragePoolId: 1
-    Name: "/{{ .Spec.Domain }}:hdd"
-    ErasureSpecies: "block-4-2"
-    VDiskKind: "Default"
-    Kind: "hdd"
-    NumGroups: 7
-    PDiskFilter {
-      Property {
-        Type: ROT
-      }
-    }
-  }
-}
-
-`
