--- internal/configuration/templates/kqp.go
@@ -1,5 +1,0 @@
-package templates
-
-const KQPConfigTemplate = `
-Enable: true
-`
