--- internal/configuration/templates/pq.go
@@ -1,5 +1,0 @@
-package templates
-
-const PQConfigTemplate = `
-Enabled: false
-`
