--- internal/configuration/templates/sys.go
@@ -1,47 +1,0 @@
-package templates
-
-const ActorSystemConfigTemplate = `
-Executor {
-  Type: BASIC
-  Threads: 9
-  SpinThreshold: 1
-  Name: "System"
-}
-Executor {
-  Type: BASIC
-  Threads: 16
-  SpinThreshold: 1
-  Name: "User"
-}
-Executor {
-  Type: BASIC
-  Threads: 7
-  SpinThreshold: 1
-  Name: "Batch"
-}
-Executor {
-  Type: IO
-  Threads: 1
-  Name: "IO"
-}
-Executor {
-  Type: BASIC
-  Threads: 3
-  SpinThreshold: 10
-  Name: "IC"
-  TimePerMailboxMicroSecs: 100
-}
-Scheduler {
-  Resolution: 64
-  SpinThreshold: 0
-  ProgressThreshold: 10000
-}
-SysExecutor: 0
-UserExecutor: 1
-IoExecutor: 3
-BatchExecutor: 2
-ServiceExecutor {
-  ServiceName: "Interconnect"
-  ExecutorId: 4
-}
-`
