--- internal/configuration/templates/vdisks.go
@@ -1,20 +1,0 @@
-package templates
-
-const VDiskConfigTemplate = `
-VDiskKinds {
-  Kind: Log
-  BaseKind: Default
-  Config {
-    FreshUseDreg: true
-    AllowKeepFlags: false
-  }
-}
-VDiskKinds {
-  Kind: LocalMode
-  BaseKind: Default
-  Config {
-    HullCompLevel0MaxSstsAtOnce: 2
-    HullCompSortedPartsNum: 1
-  }
-}
-`
